Added joi check for invites and null validation

+++ service/invite-service.js
@@ -1,5 +1,7 @@
 const uuid = require('short-uuid')
-const error = require("../object/error");
+const error = require("../object/error")
+const roleCheck = require('./roles/roleCheck')
+const postInviteSchema = require("./joi-schemas/invite-schemas").postInviteSchema
 
 class InviteService {
     constructor(dbCalendar, dbUser, dbInvite) {
@@ -13,24 +15,31 @@ class InviteService {
     }
 
     postInvite(userId, invite){
-        //TODO Joi
+        const result = postInviteSchema.validate(invite, {stripUnknown: true})
+        if (result.error)
+            return Promise.reject(error(400, result.error.details[0].message))
+        invite = Object.assign({}, result.value)
+
         return Promise.all([
             this.dbUser.getUserByEmail(invite.email),
             this.dbUser.getUser(userId),
-            this.dbCalendar.getCalendarName(invite.calendarId)
+            this.dbCalendar.getCalendarOwnerAndName(invite.calendarId)
         ])
             .then(res => {
                 const invitee = res[0]
                 const inviter = res[1]
-                const calendarName = res[2].name
+                const calendar = res[2]
 
-                //TODO Permissions
-                //TODO Check for null
+                if(!calendar || !roleCheck.isOwner(calendar, userId))
+                    return Promise.reject(error(404, 'Calendar Not Found'))
+
+                if(!invitee)
+                    return Promise.reject(error(404, 'User Not Found'))
 
                 invite.id = uuid.generate()
                 invite.inviteeId = invitee.id
                 invite.inviterId = inviter.id
-                invite.calendarName = calendarName
+                invite.calendarName = calendar.name
                 invite.inviterName = inviter.name
                 invite.inviteeName = invitee.name
                 invite.inviteeEmail = invitee.email
@@ -40,23 +49,30 @@ class InviteService {
     }
 
     getPending(userId) {
-        //TODO Permissions
-        //TODO Check for null
         return this.dbInvite.getPending(userId)
     }
 
     getSent(userId, calendarId) {
-        //TODO Permissions
-        //TODO Check for null
-        return this.dbInvite.getSent(calendarId)
+        return this.dbCalendar.getCalendarOwner(calendarId)
+            .then(calendar => {
+                if(!calendar || !roleCheck.isOwner(calendar, userId))
+                    return Promise.reject(error(404, 'Calendar Not Found'))
+
+                return this.dbInvite.getSent(calendarId)
+            })
     }
 
     acceptInvite(userId, inviteId) {
-        return this.dbInvite.deleteInvite(inviteId)
+        return this.dbInvite.getInviteeId(inviteId)
             .then(invite => {
+                if(!invite || !roleCheck.isSame(userId, invite.inviteeId))
+                    return Promise.reject(error(404, 'Invite Not Found'))
 
-                //TODO Permissions
-                //TODO Check for null
+                return this.dbInvite.deleteInvite(inviteId)
+            })
+            .then(invite => {
+                if(!invite)
+                    return Promise.reject(error(404, 'Invite Not Found'))
 
                 const participating = {
                     'calendarId': invite.calendarId,
@@ -82,19 +98,33 @@ class InviteService {
     }
 
     deleteInvite(userId, inviteId) {
-        return this.dbInvite.deleteInvite(inviteId)
+        return this.dbInvite.getInviterId(inviteId)
             .then(invite => {
-                //TODO Permissions
-                //TODO Check for null
+                if(!invite || !roleCheck.isSame(userId, invite.inviterId))
+                    return Promise.reject(error(404, 'Invite Not Found'))
+
+                return this.dbInvite.deleteInvite(inviteId)
+            })
+            .then(invite => {
+                if(!invite)
+                    return Promise.reject(error(404, 'Invite Not Found'))
+
                 return {'message': 'Deleted invite'}
             })
     }
 
     declineInvite(userId, inviteId) {
-        return this.dbInvite.deleteInvite(inviteId)
+        return this.dbInvite.getInviteeId(inviteId)
             .then(invite => {
-                //TODO Permissions
-                //TODO Check for null
+                if(!invite || !roleCheck.isSame(userId, invite.inviteeId))
+                    return Promise.reject(error(404, 'Invite Not Found'))
+
+                return this.dbInvite.deleteInvite(inviteId)
+            })
+            .then(invite => {
+                if(!invite)
+                    return Promise.reject(error(404, 'Invite Not Found'))
+
                 return {'message': 'Declined invite'}
             })
     }
